marker animator for android and ios

+++ mapbox_gl_platform_interface/lib/src/animated_marker.dart
@@ -34,7 +34,7 @@ class AnimatedRotation {
   AnimatedRotation({ this.rotation, this.duration });
 
   dynamic toJson() {
-    return <double>[rotation, duration.inMilliseconds.toDouble()];
+    return <double>[rotation.toDouble(), duration.inMilliseconds.toDouble()];
   }
 }
